requirements.txt
python-telegram-bot[job-queue]==21.10
yt-dlp
instaloader
pydantic-settings
aiofiles
ffmpeg-python
